requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0
flake8==5.0.4
flake8-bugbear==22.10.27
flake8-builtins==2.0.1
flake8-commas==2.1.0
flake8-import-order==0.18.1
flake8-variables-names==0.0.5
h11==0.14.0
idna==3.4
mccabe==0.7.0
outcome==1.2.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
